Sign exes in bob
